Ajout de validations

+++ app/Http/Controllers/IngredientController.php
@@ -6,16 +6,17 @@ use Illuminate\Http\Request;
 use Illuminate\View\View;
 use App\Http\Requests\IngredientRequest;
 use App\Models\Ingredient;
+use App\Models\Garniture;
 use Illuminate\Support\Facades\Storage;
 
 class IngredientController extends Controller
 {
     public function index(): view
     {
-        $ingredients = Ingredient::paginate(3);
+        $ingredients = Ingredient::paginate(2);
         $titre = 'Ingrédients';
         $data = ["title" => $titre, 'ingredients' => $ingredients];
-        
+
         return view('ingredient-index', $data);
     }
 
@@ -36,30 +37,43 @@ class IngredientController extends Controller
 
     public function delete(int $id)
     {
-        $ingredient = Ingredient::find($id);
-        Storage::disk('public')->delete($ingredient->picture);
-        Ingredient::destroy($id);
-        return redirect()->route('ingredients')->with('info2','ingrédient supprimée');
+        $garnitures = Garniture::where('idIngredient', $id)->get();
+        if ($garnitures->count() != 0) {
+            return redirect()->route('ingredients')->with('info2', 'Ingrédient utilisé dans une pizza');
+        } else {
+            $ingredient = Ingredient::find($id);
+            Storage::disk('public')->delete($ingredient->picture);
+            Ingredient::destroy($id);
+            return redirect()->route('ingredients')->with('info2', 'ingrédient supprimée');
+        }
     }
 
     public function save(IngredientRequest $request)
     {
+        $ingredients = Ingredient::all();
         $filename = time() . '.' . $request->picture->extension();
         $picture = $request->picture->storeAs("images", $filename, 'public');
         $ingredientModel = new Ingredient;
         if (isset($request->id)) {
             $ingredientModel = Ingredient::find($request->id);
             Storage::disk('public')->delete($ingredientModel->picture);
-            
+
             $ingredientModel->text = $request->text;
             $ingredientModel->picture = $picture;
         } else {
-
-            $ingredientModel->text = $request->text;
-            $ingredientModel->picture = $picture;
+            if ($ingredients->count() == 0) {
+                $ingredientModel->text = config('app.nom');
+                $ingredientModel->picture = config('app.picture');
+                $ingredientModel->save();
+                $ingredientModel = $request->text;
+                $ingredientModel = $request->picture;
+            } else {
+                $ingredientModel->text = $request->text;
+                $ingredientModel->picture = $picture;
+            }
         }
 
         $ingredientModel->save();
-        return redirect()->route('ingredients')->with('info','ingrédient sauvegardée');
+        return redirect()->route('ingredients')->with('info', 'ingrédient enregistrée');
     }
 }
